Create index.html
This is the html file that serves a very basic webpage that connects to the mood.sol's public methods and accesses them same as using Remix. Call on a user's Metamask wallet to connect to the page and after connection and requests approval to write to the contract the mood under set mood and read from the contract under get mood (the mood provided to the contract shows in the dev logs on the page. Also takes gas payment to set the mood from the users mood typed in the box as input.
Note: 2 items in the code are deprecated and will have to be updated at some point. You will see these items noted in the dev log in the console on the page when you run the script.